Refactor all calcprops onto parent File item
